+++ scripts/macros/2024/monsterFeatures/vampire/beguile.js
@@ -1,0 +1,6 @@
+export let beguile = {
+    name: 'Beguile',
+    monster: 'Vampire',
+    version: '1.3.126',
+    rules: 'modern'
+};