+++ scripts/macros/2024/monsterFeatures/vampire/vampireCharm.js
@@ -1,0 +1,6 @@
+export let vampireCharm = {
+    name: 'Charm',
+    monster: 'Vampire',
+    version: '1.3.126',
+    rules: 'modern'
+};