+++ scripts/macros/2024/monsterFeatures/vampire/vampireWeakness.js
@@ -1,0 +1,6 @@
+export let vampireWeakness = {
+    name: 'Vampire Weakness',
+    monster: 'Vampire',
+    version: '1.3.126',
+    rules: 'modern'
+};